feat: cc-0077 click outside works for all courts
