+++ src/pages/index.tsx
@@ -2,6 +2,10 @@ import type { NextPage } from "next";
 import HeaderLayout from "../layouts/HeaderLayout";
 import dynamic from "next/dynamic";
 import { useStoreSelector } from "@/store/hooks";
+import { useDispatch } from "react-redux";
+import { changeSelectedColor } from "@/store/reducer/courtColorSlice";
+import useOnClickOutside from "@/utils/useOnClickOutside";
+import { useRef } from "react";
 
 const ProFullCourt = dynamic(() => import("@/components/ProFullCourt"), { ssr: false });
 const FullCourt = dynamic(() => import("@/components/FullCourt"), { ssr: false });
@@ -12,14 +16,23 @@ const SmallCourt = dynamic(() => import("@/components/SmallCourt"), { ssr: false
 
 const Home: NextPage = () => {
   const { name: courtName } = useStoreSelector((state) => state.courtName);
+  const dispatch = useDispatch();
+  const ref = useRef(null); // click outside the canvas area can stop color changing
+  const handleClickOutside = () => {
+    dispatch(changeSelectedColor("none"));
+    document.body.style.cursor = "auto";
+  };
+  useOnClickOutside(ref, handleClickOutside);
   return (
     <HeaderLayout>
-      {courtName === "510 m² Pro Court (17 m × 30 m)" && <ProFullCourt />}
-      {courtName === "420 m² Full Court (15 m × 28 m)" && <FullCourt />}
-      {courtName === "210 m² Pro Half Court (15 m × 14 m)" && <ProHalfCourt />}
-      {courtName === "150 m² Half Court (15 m × 10 m)" && <HalfCourt />}
-      {courtName === "70 m² Medium Court (7 m × 10 m)" && <MediumCourt />}
-      {courtName === "45 m² Small Court (5 m × 9 m)" && <SmallCourt />}
+      <div ref={ref}>
+        {courtName === "510 m² Pro Court (17 m × 30 m)" && <ProFullCourt />}
+        {courtName === "420 m² Full Court (15 m × 28 m)" && <FullCourt />}
+        {courtName === "210 m² Pro Half Court (15 m × 14 m)" && <ProHalfCourt />}
+        {courtName === "150 m² Half Court (15 m × 10 m)" && <HalfCourt />}
+        {courtName === "70 m² Medium Court (7 m × 10 m)" && <MediumCourt />}
+        {courtName === "45 m² Small Court (5 m × 9 m)" && <SmallCourt />}
+      </div>
     </HeaderLayout>
   );
 };
