+++ tsconfig.json
@@ -30,6 +30,6 @@
     "src/styles/Chakra.ts",
     "**/*.spec.tsx",
     "globals.d.ts",
-    "src/utils/axios.ts"
-, "src/components/ProFullCourt/index.tsx"  ]
+    "src/utils/axios.ts",
+  ]
 }
